requirements.txt
dill==0.2.7.1
h5py==2.7.1
numpy==1.13.3
pandas==0.20.3
pathos==0.2.1
python-dateutil==2.6.1
scikit-learn==0.19.1
scipy==1.0.0
sklearn-deap2==0.2.1
tabulate==0.8.1
